Blockout Changes
Added some block-outs of assets to zone 2 and tested the sizing of the scientists house. Changed the terrain of zone 4, added a "beginning of game" entrance, removed trees, added specific locations for quest items with some landmarks. Was also adjusting the map around for zone 4 as i was changing things.